migrate all the eigen-related test to test_eigen.py
