+++ sympy/matrices/tests/test_commonmatrix.py
@@ -1,18 +1,18 @@
 from sympy.assumptions import Q
 from sympy.core.add import Add
 from sympy.core.function import Function
-from sympy.core.numbers import Float, I, Integer, oo, pi, Rational
+from sympy.core.numbers import , I, Integer, oo, pi, Rational
 from sympy.core.singleton import S
 from sympy.core.symbol import Symbol, symbols
 from sympy.functions.elementary.complexes import Abs
 from sympy.functions.elementary.exponential import exp
 from sympy.functions.elementary.miscellaneous import sqrt
 from sympy.functions.elementary.trigonometric import cos, sin
-from sympy.matrices.common import (ShapeError, MatrixError, NonSquareMatrixError,
+from sympy.matrices.common import (ShapeError, NonSquareMatrixError,
     _MinimalMatrix, _CastableMatrix, MatrixShaping, MatrixProperties,
     MatrixOperations, MatrixArithmetic, MatrixSpecial)
 from sympy.matrices.matrices import (MatrixDeterminant,
-    MatrixReductions, MatrixSubspaces, MatrixEigen, MatrixCalculus)
+    MatrixReductions, MatrixSubspaces, MatrixCalculus)
 from sympy.matrices import (Matrix, diag, eye,
     matrix_multiply_elementwise, ones, zeros, SparseMatrix, banded)
 from sympy.simplify.simplify import simplify
@@ -95,10 +95,6 @@ class SpecialOnlyMatrix(_MinimalMatrix, _CastableMatrix, MatrixSpecial):
 
 
 class SubspaceOnlyMatrix(_MinimalMatrix, _CastableMatrix, MatrixSubspaces):
-    pass
-
-
-class EigenOnlyMatrix(_MinimalMatrix, _CastableMatrix, MatrixEigen):
     pass
 
 
@@ -1274,155 +1270,6 @@ def test_orthogonalize():
     raises(ValueError, lambda: Matrix.orthogonalize(*vecs, rankcheck=True))
 
 
-
-# EigenOnlyMatrix tests
-def test_eigenvals():
-    M = EigenOnlyMatrix([[0, 1, 1],
-                [1, 0, 0],
-                [1, 1, 1]])
-    assert M.eigenvals() == {2*S.One: 1, -S.One: 1, S.Zero: 1}
-
-    # if we cannot factor the char poly, we raise an error
-    m = Matrix([
-        [3,  0,  0, 0, -3],
-        [0, -3, -3, 0,  3],
-        [0,  3,  0, 3,  0],
-        [0,  0,  3, 0,  3],
-        [3,  0,  0, 3,  0]])
-    raises(MatrixError, lambda: m.eigenvals())
-
-
-def test_eigenvects():
-    M = EigenOnlyMatrix([[0, 1, 1],
-                [1, 0, 0],
-                [1, 1, 1]])
-    vecs = M.eigenvects()
-    for val, mult, vec_list in vecs:
-        assert len(vec_list) == 1
-        assert M*vec_list[0] == val*vec_list[0]
-
-
-def test_left_eigenvects():
-    M = EigenOnlyMatrix([[0, 1, 1],
-                [1, 0, 0],
-                [1, 1, 1]])
-    vecs = M.left_eigenvects()
-    for val, mult, vec_list in vecs:
-        assert len(vec_list) == 1
-        assert vec_list[0]*M == val*vec_list[0]
-
-
-def test_diagonalize():
-    m = EigenOnlyMatrix(2, 2, [0, -1, 1, 0])
-    raises(MatrixError, lambda: m.diagonalize(reals_only=True))
-    P, D = m.diagonalize()
-    assert D.is_diagonal()
-    assert D == Matrix([
-                 [-I, 0],
-                 [ 0, I]])
-
-    # make sure we use floats out if floats are passed in
-    m = EigenOnlyMatrix(2, 2, [0, .5, .5, 0])
-    P, D = m.diagonalize()
-    assert all(isinstance(e, Float) for e in D.values())
-    assert all(isinstance(e, Float) for e in P.values())
-
-    _, D2 = m.diagonalize(reals_only=True)
-    assert D == D2
-
-
-def test_is_diagonalizable():
-    a, b, c = symbols('a b c')
-    m = EigenOnlyMatrix(2, 2, [a, c, c, b])
-    assert m.is_symmetric()
-    assert m.is_diagonalizable()
-    assert not EigenOnlyMatrix(2, 2, [1, 1, 0, 1]).is_diagonalizable()
-
-    m = EigenOnlyMatrix(2, 2, [0, -1, 1, 0])
-    assert m.is_diagonalizable()
-    assert not m.is_diagonalizable(reals_only=True)
-
-
-def test_jordan_form():
-    m = Matrix(3, 2, [-3, 1, -3, 20, 3, 10])
-    raises(NonSquareMatrixError, lambda: m.jordan_form())
-
-    # the next two tests test the cases where the old
-    # algorithm failed due to the fact that the block structure can
-    # *NOT* be determined  from algebraic and geometric multiplicity alone
-    # This can be seen most easily when one lets compute the J.c.f. of a matrix that
-    # is in J.c.f already.
-    m = EigenOnlyMatrix(4, 4, [2, 1, 0, 0,
-                    0, 2, 1, 0,
-                    0, 0, 2, 0,
-                    0, 0, 0, 2
-    ])
-    P, J = m.jordan_form()
-    assert m == J
-
-    m = EigenOnlyMatrix(4, 4, [2, 1, 0, 0,
-                    0, 2, 0, 0,
-                    0, 0, 2, 1,
-                    0, 0, 0, 2
-    ])
-    P, J = m.jordan_form()
-    assert m == J
-
-    A = Matrix([[ 2,  4,  1,  0],
-                [-4,  2,  0,  1],
-                [ 0,  0,  2,  4],
-                [ 0,  0, -4,  2]])
-    P, J = A.jordan_form()
-    assert simplify(P*J*P.inv()) == A
-
-    assert EigenOnlyMatrix(1, 1, [1]).jordan_form() == (
-        Matrix([1]), Matrix([1]))
-    assert EigenOnlyMatrix(1, 1, [1]).jordan_form(
-        calc_transform=False) == Matrix([1])
-
-    # make sure if we cannot factor the characteristic polynomial, we raise an error
-    m = Matrix([[3, 0, 0, 0, -3], [0, -3, -3, 0, 3], [0, 3, 0, 3, 0], [0, 0, 3, 0, 3], [3, 0, 0, 3, 0]])
-    raises(MatrixError, lambda: m.jordan_form())
-
-    # make sure that if the input has floats, the output does too
-    m = Matrix([
-        [                0.6875, 0.125 + 0.1875*sqrt(3)],
-        [0.125 + 0.1875*sqrt(3),                 0.3125]])
-    P, J = m.jordan_form()
-    assert all(isinstance(x, Float) or x == 0 for x in P)
-    assert all(isinstance(x, Float) or x == 0 for x in J)
-
-
-def test_singular_values():
-    x = Symbol('x', real=True)
-
-    A = EigenOnlyMatrix([[0, 1*I], [2, 0]])
-    # if singular values can be sorted, they should be in decreasing order
-    assert A.singular_values() == [2, 1]
-
-    A = eye(3)
-    A[1, 1] = x
-    A[2, 2] = 5
-    vals = A.singular_values()
-    # since Abs(x) cannot be sorted, test set equality
-    assert set(vals) == set([5, 1, Abs(x)])
-
-    A = EigenOnlyMatrix([[sin(x), cos(x)], [-cos(x), sin(x)]])
-    vals = [sv.trigsimp() for sv in A.singular_values()]
-    assert vals == [S.One, S.One]
-
-    A = EigenOnlyMatrix([
-        [2, 4],
-        [1, 3],
-        [0, 0],
-        [0, 0]
-        ])
-    assert A.singular_values() == \
-        [sqrt(sqrt(221) + 15), sqrt(15 - sqrt(221))]
-    assert A.T.singular_values() == \
-        [sqrt(sqrt(221) + 15), sqrt(15 - sqrt(221)), 0, 0]
-
-
 # CalculusOnlyMatrix tests
 @XFAIL
 def test_diff():
@@ -1466,10 +1313,3 @@ def test_issue_13774():
     v = [1, 1, 1]
     raises(TypeError, lambda: M*v)
     raises(TypeError, lambda: v*M)
-
-
-def test___eq__():
-    assert (EigenOnlyMatrix(
-        [[0, 1, 1],
-        [1, 0, 0],
-        [1, 1, 1]]) == {}) is False
